Fix: Delete and recreate workspace to resolve persistent sync issues
This commit provides a final, definitive fix for the missing 'Scale Barcode Settings' shortcut in the 'POS Awesome' workspace.

Previous attempts to modify the workspace via patches failed, even when the patch was confirmed to be running. This indicates that the workspace document is in a deeply "stuck" or corrupted state that prevents it from being saved correctly.

This new patch, `recreate_pos_awesome_workspace.py`, takes the most robust approach:
1.  **Deletes** the existing "POS Awesome" workspace document from the database entirely.
2.  **Recreates** the workspace from the canonical `pos_awesome.json` definition file.
3.  **Clears the cache** to ensure the new, clean workspace is displayed.

This will completely reset the workspace to its intended state, overriding any customizations or corruption and definitively resolving the issue for all users. The patch has a new, unique name to ensure it is executed by the migration tool.
